Add documentation, rename filter to esm
* Rename previous esm to esm_migration
