+++ README.md
@@ -313,7 +313,7 @@ the script.
     need to configure the tag names yourself, pass a `template_literal_tags`
     option to `convert` with an array of tag name symbols.
 
-* <a href="https://github.com/rubys/ruby2js/blob/master/lib/ruby2js/filter/esimports.rb">esimports</a>
+* <a href="https://github.com/rubys/ruby2js/blob/master/lib/ruby2js/filter/esm.rb">esm</a>
 
     Provides conversion of import and export statements for use with modern ES builders like Webpack.
